+++ HISTORY.rst
@@ -7,6 +7,7 @@ Historia zmian
 ----------------
 
 * raport optymalizujący 3N (#1131),
+* liczba N dla uczelni (#1131),
 
 202109.1080-beta1
 -----------------